+++ tests/models/__init__.py
@@ -1,0 +1,2 @@
+"""tests.models: Ensure that the model layer correctly processes data and business logic,
+adhering to the expected interface defined by the ModelProtocol."""
